stt-service/requirements.txt
fastapi
uvicorn
boto3
kafka-python
python-multipart
psycopg2-binary
